+++ app/File.php
@@ -38,7 +38,7 @@ class File extends Model
         'thumb',
         'copyright',
         'description',
-        'lasteditor',
+        'user_id',
     ];
 
     protected $appends = [
@@ -294,7 +294,7 @@ class File extends Model
 
         $file = new File();
         $file->modified = $lastModified;
-        $file->lasteditor = $user->name;
+        $file->user_id = $user->id;
         $file->mime_type = $mimeType;
         $file->name = $filename;
         $file->created = $lastModified;
@@ -414,6 +414,15 @@ class File extends Model
         return $file;
     }
 
+    public static function uploadAvatar($file, $user) {
+        Storage::delete($user->avatar);
+        $filename = $user->id . "." . $file->getClientOriginalExtension();
+        return $file->storeAs(
+            'avatars',
+            $filename
+        );
+    }
+
     public function setContent($fileObject) {
         $filehandle = fopen($fileObject->getRealPath(), 'r');
         Storage::put(
@@ -448,7 +457,7 @@ class File extends Model
         $link = new EntityFile();
         $link->file_id = $this->id;
         $link->entity_id = $eid;
-        $link->lasteditor = $user->name;
+        $link->user_id = $user->id;
         $link->save();
     }
 
@@ -713,6 +722,10 @@ class File extends Model
 
     public function getExifAttribute() {
         return $this->getExifData();
+    }
+
+    public function user() {
+        return $this->belongsTo('App\User');
     }
 
     public function entities() {