+++ app/Reference.php
@@ -17,7 +17,7 @@ class Reference extends Model
         'attribute_id',
         'bibliography_id',
         'description',
-        'lasteditor',
+        'user_id',
     ];
 
     const rules = [
@@ -34,7 +34,7 @@ class Reference extends Model
         foreach($values as $k => $v) {
             $reference->{$k} = $v;
         }
-        $reference->lasteditor = $user->name;
+        $reference->user_id = $user->id;
         $reference->save();
 
         return self::with('bibliography')->find($reference->id);
@@ -45,6 +45,10 @@ class Reference extends Model
             $this->{$k} = $v;
         }
         $this->save();
+    }
+
+    public function user() {
+        return $this->belongsTo('App\User');
     }
 
     public function entity() {