+++ app/ThLanguage.php
@@ -13,10 +13,14 @@ class ThLanguage extends Model
      * @var array
      */
     protected $fillable = [
-        'lasteditor',
+        'user_id',
         'display_name',
         'short_name',
     ];
+
+    public function user() {
+        return $this->belongsTo('App\User');
+    }
 
     public function labels() {
         return $this->hasMany('App\ThConceptLabel', 'language_id');