+++ resources/js/app.js
@@ -5,6 +5,7 @@ import {
     faGithub,
     faHtml5,
     faLaravel,
+    faOrcid,
     faVuejs
 } from '@fortawesome/free-brands-svg-icons';
 import {
@@ -71,6 +72,7 @@ import {
     faFileWord,
     faFolder,
     faGlobeAfrica,
+    faIdBadge,
     faInfoCircle,
     faLayerGroup,
     faLightbulb,
@@ -82,6 +84,7 @@ import {
     faMapMarkedAlt,
     faMapMarkerAlt,
     faMicrochip,
+    faMobileAlt,
     faMonument,
     faPalette,
     faPaperPlane,
@@ -128,8 +131,11 @@ import {
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserCheck,
+    faUserCog,
     faUserEdit,
     faUsers,
+    faUserTimes,
     faVolumeMute,
     faVolumeUp
 } from '@fortawesome/free-solid-svg-icons';
@@ -158,6 +164,7 @@ import Users from './components/Users.vue';
 import Roles from './components/Roles.vue';
 import Preferences from './components/Preferences.vue';
 import UserPreferences from './components/UserPreferences.vue';
+import UserProfile from './components/UserProfile.vue';
 const DataModel = () => import(/* webpackChunkName: "group-bib" */ './components/DataModel.vue')
 const DataModelDetailView = () => import(/* webpackChunkName: "group-bib" */ './components/DataModelDetailView.vue')
 
@@ -238,6 +245,7 @@ library.add(
     faFileWord,
     faFolder,
     faGlobeAfrica,
+    faIdBadge,
     faInfoCircle,
     faLayerGroup,
     faLightbulb,
@@ -249,7 +257,9 @@ library.add(
     faMapMarkedAlt,
     faMapMarkerAlt,
     faMicrochip,
+    faMobileAlt,
     faMonument,
+    faOrcid,
     faPalette,
     faPaperPlane,
     faPause,
@@ -296,7 +306,10 @@ library.add(
     faUnlink,
     faUnlockAlt,
     faUser,
+    faUserCheck,
+    faUserCog,
     faUserEdit,
+    faUserTimes,
     faUsers,
     faVolumeMute,
     faVolumeUp
@@ -328,10 +341,14 @@ require('popper.js');
 require('bootstrap');
 window.Vue = require('vue');
 window._clone = require('lodash/clone');
+window._cloneDeep = require('lodash/cloneDeep');
 window._orderBy = require('lodash/orderBy');
 window._debounce = require('lodash/debounce');
 $ = jQuery  = window.$ = window.jQuery = require('jquery');
 require('./globals.js');
+
+let utc = require('dayjs/plugin/utc')
+dayjs.extend(utc);
 
 // Create Axios instance for external (API) calls
 Vue.prototype.$externalHttp = Axios.create({
@@ -545,6 +562,14 @@ const router = new VueRouter({
                 auth: true
             }
         },
+        {
+            path: '/profile',
+            name: 'userprofile',
+            component: UserProfile,
+            meta: {
+                auth: true
+            }
+        },
     ]
 });
 // Workaround to load plugin pages, whose routes
@@ -627,6 +652,7 @@ import AttributeSearch from './components/AttributeSearch.vue';
 import CsvTable from './components/CsvTable.vue';
 
 // Reusable Components
+import UserAvatar from './components/UserAvatar.vue';
 import Attributes from './components/AttributeList.vue';
 import EntityTree from './components/EntityTree.vue';
 import EntityTypes from './components/EntityTypeList.vue';
@@ -646,6 +672,7 @@ Vue.component('entity-type-search', EntityTypeSearch);
 Vue.component('label-search', LabelSearch);
 Vue.component('attribute-search', AttributeSearch);
 Vue.component('csv-table', CsvTable);
+Vue.component('user-avatar', UserAvatar);
 Vue.component('attributes', Attributes);
 Vue.component('entity-tree', EntityTree);
 Vue.component('entity-types', EntityTypes);