+++ resources/js/i18n/en.js
@@ -2,6 +2,8 @@ const en = {
     global: {
         save: 'Save',
         delete: 'Delete',
+        deactivate: 'Deactivate',
+        reactivate: 'Reactivate',
         remove: 'Remove',
         cancel: 'Cancel',
         close: 'Close',
@@ -20,6 +22,11 @@ const en = {
         'delete-name': {
             title: 'Delete {name}',
             desc: 'Do you really want to delete {name}?'
+        },
+        'deactivate-name': {
+            title: 'Deactivate {name}',
+            desc: 'Do you really want to deactivate {name}?',
+            info: 'Certain entries can not be deleted or have to be deactivated first. Otherwise deleting these entries could lead to data loss.'
         },
         'edit-name': {
             title: 'Edit {name}'
@@ -74,7 +81,14 @@ const en = {
         },
         user: {
             settings: 'Preferences',
-            logout: 'Logout'
+            profile: 'Profile',
+            logout: 'Logout',
+            info_title: 'User Information',
+            member_since: 'Member since',
+            deactivated_since: '<span class="font-weight-bold">deactivated</span> since {dt}',
+            contact: 'Contact',
+            avatar: 'Avatar',
+            invalid_orcid: 'This ORCID is invalid',
         },
         select: {
             placehoder: 'Select option',
@@ -130,14 +144,18 @@ const en = {
         'display-name': 'Displayname',
         email: 'E-Mail Address',
         email_or_nick: 'E-Mail Address or Nickname',
+        phonenumber: 'Phone number',
         password: 'Password',
         'remember-me': 'Remember me',
+        'orcid': 'ORCID',
         description: 'Description',
         roles: 'Roles',
         permissions: 'Permissions',
         'added-at': 'Added',
         'created-at': 'Created',
         'updated-at': 'Updated',
+        'deleted-at': 'Deleted',
+        'deactivated-at': 'Deactivated',
         options: 'Options',
         type: 'Type',
         'root-attribute': 'Parent-Attribute',
@@ -246,6 +264,9 @@ const en = {
         },
         user: {
             'add-button': 'Add new User',
+            active_users: 'Active Users',
+            deactivated_users: 'Deactivated Users',
+            empty_list: 'User-list is empty',
             toasts: {
                 updated: {
                     title: 'User updated',