+++ tests/Feature/ApiBibliographyTest.php
@@ -66,7 +66,7 @@ class ApiBibliographyTest extends TestCase
             'organization',
             'school',
             'series',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at'
         ]);
@@ -154,7 +154,7 @@ class ApiBibliographyTest extends TestCase
                 'organization',
                 'school',
                 'series',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at',
             ]
@@ -296,7 +296,7 @@ class ApiBibliographyTest extends TestCase
             'organization',
             'school',
             'series',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at'
         ]);