+++ tests/Feature/ApiFileTest.php
@@ -72,7 +72,7 @@ class ApiFileTest extends TestCase
             'copyright',
             'description',
             'mime_type',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'url',
@@ -97,7 +97,7 @@ class ApiFileTest extends TestCase
             'copyright' => null,
             'description' => null,
             'mime_type' => 'image/png',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'created_at' => '2019-03-08 13:13:11',
             'updated_at' => '2019-03-08 13:13:11',
             'url' => "/storage/spacialist_screenshot.png",
@@ -137,7 +137,7 @@ class ApiFileTest extends TestCase
             'copyright',
             'description',
             'mime_type',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'url',
@@ -164,7 +164,7 @@ class ApiFileTest extends TestCase
             'copyright' => 'Vinzenz Rosenkranz (CC BY-NC-SA 2.0)',
             'description' => 'Edinburgh Castle',
             'mime_type' => 'image/jpeg',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'created_at' => '2019-03-08 13:13:11',
             'updated_at' => '2019-03-08 13:13:12',
             'url' => "/storage/test_img_edin.jpg",
@@ -338,7 +338,7 @@ class ApiFileTest extends TestCase
                 'copyright' => null,
                 'description' => null,
                 'mime_type' => 'text/plain',
-                'lasteditor' => 'Admin',
+                'user_id' => 1,
                 'created_at' => '2019-03-08 13:13:11',
                 'updated_at' => '2019-03-08 13:13:11',
                 'category' => 'text',
@@ -371,7 +371,7 @@ class ApiFileTest extends TestCase
                 'copyright',
                 'description',
                 'mime_type',
-                'lasteditor',
+                'user_id',
                 'created_at',
                 'updated_at',
                 'category',
@@ -528,7 +528,7 @@ class ApiFileTest extends TestCase
             'copyright',
             'description',
             'mime_type',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
             'category'
@@ -545,7 +545,7 @@ class ApiFileTest extends TestCase
             'copyright' => $cr,
             'description' => $desc,
             'mime_type' => 'image/png',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'created_at' => "$uplFile->created_at",
             'updated_at' => "$uplFile->updated_at",
             'category' => 'image'
@@ -587,7 +587,7 @@ class ApiFileTest extends TestCase
             'exif' => [],
             'thumb' => 'test_img_edin_thumb.0.jpg',
             'mime_type' => 'image/jpeg',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'category' => 'image'
         ]);
         $content = $response->decodeResponseJson();
@@ -626,7 +626,7 @@ class ApiFileTest extends TestCase
             'name' => 'text1.3.txt',
             'exif' => null,
             'mime_type' => 'text/plain',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
             'category' => 'text'
         ]);
         $this->assertTrue($uplFile->isText());
@@ -981,7 +981,7 @@ class ApiFileTest extends TestCase
             'copyright',
             'description',
             'mime_type',
-            'lasteditor',
+            'user_id',
             'created_at',
             'updated_at',
         ]);
@@ -993,7 +993,7 @@ class ApiFileTest extends TestCase
             'copyright' => null,
             'description' => null,
             'mime_type' => 'text/plain',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
         ]);
 
         $this->refreshToken($response);
@@ -1018,7 +1018,7 @@ class ApiFileTest extends TestCase
             'copyright' => null,
             'description' => null,
             'mime_type' => 'image/jpeg',
-            'lasteditor' => 'Admin',
+            'user_id' => 1,
         ]);
     }