+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## 0.8 - Helgö (Unreleased)
 ### Added
+- User profile to allow users to edit their information
+  - Support for phone numbers and avatars
 ### Fixed
 ### Changed
 - Switch from momentjs to dayjs for dates